backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
chromadb==0.4.24
pydantic==2.10.3
python-multipart==0.0.17
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.1.0

